navigation to MovieDetail

+++ app/src/main/java/com/example/moviesapp/data/model/Movie.kt
@@ -1,7 +1,5 @@
 package com.example.moviesapp.data.model
 
-import com.google.gson.annotations.SerializedName
-
 data class Movie(
     val id: Int = -1,
     // @SerializedName("adult")
@@ -17,7 +15,7 @@ data class Movie(
     val title: String = "",
     val video: Boolean = false,
     val vote_average: Double = -1.0,
-    val vote_count: Long = -1
+    val vote_count: Int = -1
 )
 
 data class MovieList(val results: List<Movie> = listOf())
